Add a new subspec to support OAuth-only integrations
This is useful for integrations with robust server-side components and
relatively dumb, untrusted clients that do not hold the client secret.
