+++ Pod/Classes/CoinbaseDefines.m
@@ -1,0 +1,4 @@
+#import "CoinbaseDefines.h"
+
+NSString *const CoinbaseErrorDomain = @"CoinbaseErrorDomain";
+
